fix(config) parse headers config option from .kong_env
This fixes a bug introduced in 976dd872 (#3300)
where headers config option was being overwritten
as a table. The table is not persisted in the intermediate
config file and hence the config option does not take effect.

This commit also exposes `get_running_conf` helper function to
read conf from prefix directory in use during the current test.